# api/openapi.yaml
openapi: 3.0.3
info:
  title: Golf Course Reservation Management System API
  description: 골프장 예약 관리 시스템 API
  version: 1.0.0
  contact:
    name: API Support
    email: support@golfadmin.com
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.golfadmin.com
    description: Production server
security:
  - bearerAuth: []
...
paths:
...
  /performance/complete:
    post:
      tags:
        - Performance
      summary: 티타임 완료 등록
      operationId: completePerformance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompletePerformanceRequest'
      responses:
        '201':
          description: 완료 등록 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Performance'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /performance/summary:
    get:
      tags:
        - Performance
      summary: 실적 요약 조회
      operationId: getPerformanceSummary
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: managerId
          in: query
          schema:
            type: string
        - name: teamId
          in: query
          schema:
            type: string
        - name: golfCourseId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 실적 요약
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerformanceSummaryResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /performance/stats:
    get:
      tags:
        - Performance
      summary: 실적 통계 조회
      operationId: getPerformanceStats
      parameters:
        - name: period
          in: query
          required: true
          schema:
            type: string
            enum: [daily, weekly, monthly, yearly]
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [manager, team, golfCourse, region]
      responses:
        '200':
          description: 실적 통계
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerformanceStatsResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  # Golf Courses Endpoints
  /golf-courses:
    get:
      tags:
        - GolfCourses
      summary: 골프장 목록 조회
      operationId: getGolfCoursesList
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SortParam'
        - name: region
          in: query
          schema:
            $ref: '#/components/schemas/Region'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/GolfCourseStatus'
      responses:
        '200':
          description: 골프장 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GolfCourseListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - GolfCourses
      summary: 골프장 등록
      operationId: createGolfCourse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGolfCourseRequest'
      responses:
        '201':
          description: 골프장 등록 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GolfCourse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /golf-courses/{id}:
    get:
      tags:
        - GolfCourses
      summary: 골프장 상세 조회
      operationId: getGolfCourseById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 골프장 상세 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GolfCourse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      tags:
        - GolfCourses
      summary: 골프장 수정
      operationId: updateGolfCourse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGolfCourseRequest'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GolfCourse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags:
        - GolfCourses
      summary: 골프장 삭제
      operationId: deleteGolfCourse
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: 삭제 성공
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # Users Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      description: 페이지 번호 (1부터 시작)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: 페이지당 항목 수
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      description: 정렬 기준 (field:asc 또는 field:desc)
      schema:
        type: string
        pattern: '^[a-zA-Z]+:(asc|desc)$'
  schemas:
...
    AccountType:
      type: string
      enum:
        - SUPER_ADMIN
        - ADMIN
        - TEAM_LEADER
        - INTERNAL_MANAGER
        - EXTERNAL_MANAGER
        - PARTNER
        - GOLF_COURSE
        - MEMBER
    Region:
      type: string
      enum:
        - 제주
        - 경기
        - 강원
        - 충청
        - 호남
        - 영남
        - 경상
        - 기타
    TeeTimeStatus:
      type: string
      enum:
        - AVAILABLE
        - RESERVED
        - CONFIRMED
        - COMPLETED
        - CANCELLED
    TeeTimeType:
      type: string
      enum:
        - DAILY
        - PACKAGE
    BookingType:
      type: string
      enum:
        - BOOKING
        - JOIN
    TimeSlot:
      type: string
      enum:
        - 1부
        - 2부
        - 3부
    GolfCourseStatus:
      type: string
      enum:
        - ACTIVE
        - API_INTEGRATION
        - MANUAL
        - INACTIVE

    # Error Response
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: 에러 코드
        message:
          type: string
          description: 에러 메시지
        details:
          type: object
          description: 추가 에러 상세 정보
    ValidationError:
      type: object
      required:
        - code
        - message
        - errors
      properties:
        code:
          type: string
          enum: [VALIDATION_ERROR]
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              code:
                type: string
...
    User:
      type: object
      properties:
        id:
          type: string
        phoneNumber:
          type: string
        name:
          type: string
        accountType:
          $ref: '#/components/schemas/AccountType'
        teamId:
          type: string
          nullable: true
        team:
          $ref: '#/components/schemas/Team'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Team:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        leaderId:
          type: string
        leader:
          $ref: '#/components/schemas/User'
        members:
          type: array
          items:
            $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    GolfCourse:
      type: object
      properties:
        id:
          type: string
        order:
          type: integer
        region:
          $ref: '#/components/schemas/Region'
        name:
          type: string
        address:
          type: string
        phoneNumber:
          type: string
        status:
          $ref: '#/components/schemas/GolfCourseStatus'
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateGolfCourseRequest:
      type: object
      required:
        - region
        - name
        - address
        - phoneNumber
      properties:
        order:
          type: integer
        region:
          $ref: '#/components/schemas/Region'
        name:
          type: string
        address:
          type: string
        phoneNumber:
          type: string
        status:
          $ref: '#/components/schemas/GolfCourseStatus'
          default: ACTIVE
        notes:
          type: string
    UpdateGolfCourseRequest:
      type: object
      properties:
        order:
          type: integer
        region:
          $ref: '#/components/schemas/Region'
        name:
          type: string
        address:
          type: string
        phoneNumber:
          type: string
        status:
          $ref: '#/components/schemas/GolfCourseStatus'
        notes:
          type: string
    GolfCourseListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/GolfCourse'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Tee Time Schemas
    TeeTime:
      type: object
      properties:
        id:
          type: string
        golfCourseId:
          type: string
        golfCourse:
          $ref: '#/components/schemas/GolfCourse'
        date:
          type: string
          format: date
        time:
          type: string
          pattern: '^([01][0-9]|2[0-3]):[0-5][0-9]$'
        timeSlot:
          $ref: '#/components/schemas/TimeSlot'
        greenFee:
          type: number
          format: decimal
          description: 그린피 (만원 단위, 소수점 1자리)
        playerCount:
          type: integer
          minimum: 1
          maximum: 4
        requirements:
          type: string
        holeType:
          type: string
          enum: [18홀, 9홀]
        caddie:
          type: boolean
        advance:
          type: number
          format: decimal
          description: 선입금
        mealIncluded:
          type: boolean
        cartIncluded:
          type: boolean
        teeTimeType:
          $ref: '#/components/schemas/TeeTimeType'
        bookingType:
          $ref: '#/components/schemas/BookingType'
        status:
          $ref: '#/components/schemas/TeeTimeStatus'
        reservedAt:
          type: string
          format: date-time
          nullable: true
        managerId:
          type: string
          nullable: true
        manager:
          $ref: '#/components/schemas/User'
        confirmedAt:
          type: string
          format: date-time
          nullable: true
        confirmedBy:
          type: string
          nullable: true
        connectedTeeTimeId:
          type: string
          nullable: true
          description: 연결된 티타임 ID (패키지용)
        accommodationInfo:
          type: string
          nullable: true
          description: 숙박 정보 (패키지용)
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Performance:
      type: object
      properties:
        id:
          type: string
        teeTimeId:
          type: string
        teeTime:
          $ref: '#/components/schemas/TeeTime'
        completedAt:
          type: string
          format: date-time
        actualPlayerCount:
          type: integer
        actualGreenFee:
          type: number
          format: decimal
        revenue:
          type: number
          format: decimal
        notes:
          type: string
        registeredBy:
          type: string
        registeredByUser:
          $ref: '#/components/schemas/User'
        createdAt:
          type: string
          format: date-time
    CompletePerformanceRequest:
      type: object
      required:
        - teeTimeId
        - actualPlayerCount
        - actualGreenFee
      properties:
        teeTimeId:
          type: string
        actualPlayerCount:
          type: integer
          minimum: 1
          maximum: 4
        actualGreenFee:
          type: number
          format: decimal
        notes:
          type: string
    PerformanceSummaryResponse:
      type: object
      properties:
        totalRevenue:
          type: number
          format: decimal
        totalCount:
          type: integer
        averageGreenFee:
          type: number
          format: decimal
        performances:
          type: array
          items:
            $ref: '#/components/schemas/Performance'
    PerformanceStatsResponse:
      type: object
      properties:
        period:
          type: string
        stats:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              revenue:
                type: number
                format: decimal
              count:
                type: integer
              averageGreenFee:
                type: number
                format: decimal

    # Common Schemas
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
  responses:
    BadRequestError:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
    UnauthorizedError:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: 인증이 필요합니다
    ForbiddenError:
      description: 권한 부족
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FORBIDDEN
            message: 해당 작업을 수행할 권한이 없습니다
    NotFoundError:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: NOT_FOUND
            message: 요청한 리소스를 찾을 수 없습니다
...